Propositional logic types, lexing, parsing
Following the book formula is generic over some atomic type.
And, for propositional logic the atomic type is a `Prop`.

In contrast to the book formulas are enumerated by operator.
This is (somewhat) useful for abstraction when matching, etc.

The generic parsing from the book is replaced with simple recursive descent, similar to arithmetic.
Further, the lexer is extended to support various ways of expressing constants.

In another break from the book printing formulas makes use of unicode.