Broad generalisation from PropFormula to Formula
And, in addition, generalised constructors for formulas.

In short, existing PropFormula constructors are now Formula<T> constructors.
And, the `Prop` struct has been moved to a separate file for clarity.
The additional constructors take ops, etc., and along with generalising the existing constructors indirectly call the additional (in order to keep a canonical constructor path) --- the fn calls should (hopefully) be inlined.