on_atoms fn
Following the book, mostly.

The pure implementation requires (at least for this simple implementation) Formulas to implement Clone.
As it's highly unlikely there will be any interest in formulas which don't indirectly implement Clone, this has been added as a bound of atoms.

A mutating variant differs slightly, to support mutation, but otherwise follows the same pattern.