Valuations, and evaluations
Also, `atom_union` to coincide with the book.

Valuations are handled as hash maps for simplicity.
Though if this raises some issues it may be revised.
A map here is used by default as it corresponds to a partial fn.

Ideally, some error handling would be implemented, though things aren't ideal...