Valuation permutations, etc.
The book implements a recursive function to inspect all valuations.
This is elegant, though we take a different approach.
In place of recursively generating all permutations a method to generate the next permutation of a valuation is implemented.
With this, inspecting all valuations is implemented through generating a default valuation and permuting the valuation some number of times.

The slight headache is the structure of valuations.
Fast lookup is important, but some fixed order is required for permutations.
As far as I can tell, Rust doesn't guarantee iterators over a fixed hash set is of fixed order.
So, valuations are split into two parts.
- A vector of bools
- A map from props to indices of the vector.
Lookups no require a diversion, though almost all interactions remain quite simple.