Truth tables
In place of test an example is included.

The implementation differs a little from the book as permutations are used and in addition some work is required to recover the order of atoms.
The results, though, are mostly equivalent.
Though, order is arbitrary as the valuation is built from a hash set.