Simplification, primarily, and some tidying
Simplification follows the book, and slightly improves on the code for arithmetic.
While typing things up the (mis)use of the `Op` prefix on formulas was noted, and has been removed.

As the previous collection of methods (mostly) focused on atoms, these have been moved to a new file.
And, in addtion, modules have been tidied, with, e.g., only the top parsing fn having a public export.