literals and negation
The book assumes negative and positive are only applied to literals.
This is fixed here, and I guess omitted in the book for space considerations.
As, I guess, there's no easy way to match against a string in OCaml.

For clarity the methods have been given more descriptive names.

Negation, also, is generalised to apply to any formula.
Though, quantifiers are todo (as one could push a negation or pop a negation from the dual).