Formula sets, with some DNF impls
In broad structure this follows the book by representing sets as lists / vecs.

The implementation is a little more structured, though, as sets only contain literals.
Where, here, literals are (Atomic, bool) pairs.
It's a little more work to track the bool, but the trade is worthwhile to enforce the literal constraint.

Further, the sets include a mode to highlight whether the intended interpretation is CNF or DNF.
And, tautologies and contradictions are apropriately handled.

As these extensions are not present in the book, it's unlikely they'll have any use.