Subsumption, and explicit comparisons
Subsumption is simple, so long as the ordering of sets is fixed to allow comparisons between atoms.
To ensure this is the case, custom sorting fns have been implemented.
These ensure the vecs are sorted by the literals contained, rather than length, and so support tests of inclusion by comparing adjacent vecs.

In addition, typedefs have been included for literals and sets of literals.

Literlas could be revised to a struct, and then cmp could be impl'd...