Initial prop dictionary structure
To be used for definitional CNF.

The idea is to store a definition generator and a map from formulas to definitions.
Insertion, then, amounts to generating fresh definitions until a definition is found which is not part of the formula to be defined.

For the formulas must be hashable, but otherwise the additions are local.

Small tests included.