BDD foundations
Possibly functional.

The OCaml isn't particularly easy to parse, there's a lack of examples in the book, though working to intuition on how things should be...

The issue is negation.
Specifically, negation and extant nodes.
The approach taken is to always store nodes by absolute value, and to swap branches if the node is negated...