General tidying
Parsing has been moved to a separate module, in anticipation of FOL.
A `parse_propositional` fn is present at the top level of the module, and the same fn is aliased as `parse` within the propositional mod, to allow contextual shortening when relevant.

Atoms have been revised to wrap the underlying atomic, rather that storing the atomic as a struct field.
As, atoms (will / should) not require any additional information.

Atomic now requires implementation of an `id` method, which in turn allows definition of the same for literals.

The prop `name` field is now `id`