Simple parsing for terms and relations
Both parsers try to parse a term / relation, and only mutate a stream of tokens if an identifier is found.
So, sufficient information that a term / relation follows or else the input is malformed.
Neither is particularly elegant, but both seem to function ok.

For terms and relations, commas are now lexed.
At the moment there's no use for these, so perhaps they should be omitted...

Elsewhere, the `Constant` token has been renamed to `Identifier`, for clarity.