Term const / var parsing refinments
Things should mostly be in line with the book now.

Parsing FOL is yet to be implemented, though the expectation is to keep a store of all variable ids, adding to the store when parsing a quanifier, and using this to resovle whether or not to classify some id as a variable.