First order connective parsing
Mostly, mostly, the same as propositional logic.

Minor differences are a persistent set of variables in scope, and some changes to parsing atoms / relations.
Here, as parsing relations requires an identifier, the match is set to only advance when anything but an identifier matches.

Light comments added on the above.