Initial pass at quantifier parsing
Formulas have been updated to store terms in quantifier variable places.
It would be nice to have a clear separation between propositional and first order.
In practice, though, the cost of repeat code is too high, and using atoms as before likewise makes a lot of work.
As the propositional code ignores quantifiers, though, this has no cascading effects.

Quantifier parsing is mostly straightforward, with quantifiers handed similarly to negation.
The only issue is (potentally) correcting a quantified constant.
For what I've read of the book this seems intended.
And, otherwise, similar code should be preserved to ensure variables are variables.

To help clarify variables and constants ansi escape codes have been introduced when displaying.