Simplified quantifier parsing
Quantifier tokens unified with a quantifer enum to break the cases.

The full stop is optional, for now.
I guess it may be useful for collecting multiple variables, though for now only a single variable is supported.

A `to_variable` method now allows casting a constant to a variable.

As with all the code so far, there are many paths to panicing.
Eventually errors may be handled...