More lexing revisions
Validation of an identifier is now part of `string_to_token`.
To make this work creating of a string must break at delimiters, so a string of delimiters is declared as static.

panics return to lexing, as I'm not quite so sure about how I want to handle some possible identifiers.

Some steps forward and back, I think...