Checkpoint of first order evaluation
Following the book, roughly.

Terms are now variables or functions, with constants defined as zero argument terms.
In addition, variables and functions are independent structs, with the term enum wrapping these.
As, with this:

A valuation is a (hash) map from variables to elements of some domain.
An interpretation is a function form a function and valuation to elements of some domain.

Interpretations diverge from the book by requiring a valuation and returning elements from the domain.
Still, this simplifies evaluation, as the interpretation can be used to evaluate the function, given an evaluation.
That is, there's no application in first order logic, so evaluating a function amounts to applying the interpreted function to the evaluted / interpreted arguments.
Further, there's no point in inspecting functions without a valuation.
And, in Rust there would be significant work required to return generic functions.

Elsewhere, display of terms has been rewritten to avoid allocations, and functions are now given in bold.