Foundations of quantifier evaluation
The domain trait was almost the right idea, I think.

The right idea, I hope, is to require the relevant methods are generic over elements of the domain, and to include the domain as a structre, as part of the interpretation (rather than as a parameter).
With this construction of domains is simple, and the relevant bounds are simple to enforce, as before.

Crude evaluation of quantifiers is implemented.

At the moment there is no shadowing of variables, nor requirements that the domain is non-empty.
Still, these could be fixed with roughly as many words as this commit...