Term tidying
Parsing now requires a variable to follow a quantifier.
This marks the to_variable method on terms as obsolete, and it has been removed.
This restriction makes almost everything easier, and I'll be (very) surprised if the book makes use of binding constants.

Construction methods have been capitalised, to match existing convention.

There's now a single constructor for Fun terms, at the expense of some clones which I'd guess are good candidates for optimisation.

And, `is_const_id` has been moved to parsing, as it's only relevant when parsing.