First order substitution, first pass
Replication of the book was attempted.
Speficially, by modifying the substitution through closures to override existing substitutions.
The issue, though, is that closures have unique types, and substitution is recursive, so type resolution could not be completed.

The alternative implemented is to carry a map of ad-hoc overrides.
This, I think, is a reasonable trade-off as the map is used only for variations, and these are limited to direct mappings.

Of note, the map is hidden to any observer outside the syntax module.