Generalised quantification
... and simplification.

Some significant changes.

Ideally, simplification and other methods are with respect to formulas.
This was an issue, as to inspect, e.g., the variables found in an atom one (ideally) needs a trait method which is generic over atoms.

This is what's been implemented.
The atomic trait holds a quantum type --- the thing quantified over --- and requires a variable method on atoms.
With this, booleans have an 'never' quantum, while relations have a var.
To help with generality the variable method returns an iterator, but it's a fairly simple definition for relations.

Some fns / methods required replacing a var argument with quantum, but as relevant cases were already generic over atomic this did not lead to any issues.

Perhaps hindsight won't be so kind, but this seems ok as a solution...