Primarily substitution updates
It seems it'll be useful to have the option to add interrupts to any substitution.
So, substitutions are now structs, generated from substitution functions, and with built in capability to handle interrupts as before.
This has the benifit of unify the interrupt code, etc.

In principle substitutions could be generalised further, and defined on formulas.
Though, for this moment I don't think it'd be worthwhile.

With this some substitution methods have been simplified / aides promoted.

Elsewhere, Formula::Quantifer has been renamed to Formula::Quantified, for improved grammar and to avoid namespace clashes if both Formulas and ops are imported.