Pulling quantifiers
Most file modifications are from renaming `expr` to `fm` in Quantified formulas.

Though, this also fixes a mistake in free variables.
There, a persisent set passed around meant the variables were always free only with respect to one branch of a binary formula.
So, the update now uses recursive set merging.

With this, pulling formulas works ok.

The case matching is very messy, and the analogue to `pullq` is inline.
Though, of note, the mess is due to deconstructing formulas, rather than regenerating formulas, so outlining `pullq` would not help much...