Atomic updates
Atomic now requires variable and function type specification, and an iterator over all functions in the atom.
The variable type replaces the quantum type.

The variable (and fresh function) iterators now iterate over all instances of the item.
This seems preferable for a generic method, and avoids allocating a set / searching.