Ids and variants
Variants are something of a headache in the book.
As, identifying (and so creating) variants relies on string inspection.

The idea is to simplify most things by storing a variant field with a Fun / Var.

This still requires some way to identify a variant from a string.
Though, this is mostly easy.
As, when reading a string a final underscore followed by digits on a non-empty prefix may specify a variant, by convention established here.

Parsing of the above is implemented and traced though construction functions.