Variants of functions
After something of a detour.

The key feature here (aside from the new variant fields) is the use of definitional equivalence between functions in order to avoid generating variants from all applications of a function.

Also, fixed some significant mistakes when determining the minimal variant.
The mistakes were overlooked, and initial tests were lucky / insufficient as an issue arises only when variants of the same are inspected out of order.