Skolemization
With the ground work, things are mostly straightforward.

A small fix for variants, as including any '_'s not used as a variant separator was overlooked.
Still, enumerating the split and adding a check fixes this.

Most of the matching for skolemization uses clones without much thought, as I expect to (hopefully) take another look at matches sometime.