Initial separation of ANSI formatting from display
ANSI formatting is nice for inspection, though it'd also be nice to have the option to (re)parse strings obtained from formulas using standard methods.
So, a fresh `ansi_string` method is introduced for the existing display impls which use ANSI.

For the moment, relations and terms are covered.

Display strings and ANSI strings are similar, though I think there's good reason against sharing code.
Specifically, as, though markup can easily be included as a flag, the methods also differ on things such as parentheses for constants.