Formula display ANSI separation
In contrast to terms, unification for formulas is clearly worthwhile, as no ANSI is handled explicitly, and any any potential ambiguity from omitted parentheses must be handled by parsing.

To make this work, then, a variant of fmt is defined with an argument for whether to use ansi, and formatting is as before, though with some extra steps to make recursive calls to the fresh fn.

In addition, the same method is required by the atomic trait, and so base cases are handled uniformly.