Custom Eq, Ord, Hash, for Funs
Motivated by initial work on Herbrand's theorem.

In short, it seems much easier to detour through other techniques than hashing when argument (in)equality matters for functions.
So, custom implementations which ignore argument (in)equality are given for Eq and Hash (and Partial variants).

In addition, ording of Funs is now by arity, id, variant, and finally by argument instances.
This may have some use, or may be revised...