Ground basics
A little different to the book, in terms of implementation.

The idea, roughly, is the same.

Generate ground by applying each function to the existing ground.

Here a struct is used for efficiency.

itertools is used for k permutations.
Motivation is interest in implementing ground, and in general itertools is likely to save further headaches and cycles.

Small tests for ground included.

In addition, fmt_ansi has been extended to Funs and Terms, to help pretty print ground.
Of note, ANSI has changed a little, with functions using italics with non-zero arity.