CNF Sets, revised one literal rule
CNF sets, simply.

The direct translation assumes a formula in CNF, and is the dual of the direct DNF translation.

For general formulas, the three tseytin -> CNF set is wrapped in a method on PropFormulas.

The one literal rule has been extended.
As, I forgot to handle removing sets containing the literal.
Though, revised, rather than (purely) extended, as the inspection loop is rewritten, esp. as there's no interest in negating literals (as both polarities on a matching atom matter).