Initial pass at the affirmative / negative rule
The implementation is given with the intent of making few passses over the formula.

Though, it may be the fewest.

As, it seems at least two passes are required.

One non-mutating pass is required, as of interset are those literals appearing with exactly one polarity, which requires inspection of each set.
Mutation can't occur here, as one can't be certain a literal will not appear in any set which is yet to be inspected.

And, then, a second mutating pass is required to remove sets containing an affirmative / negative literal.

Though, no argument that the rest of the implementation is efficient.