Formula sets and atoms
It's not so simple...

Though, two rules *are* simple.

Speficially, both affirmative / negative and resolution rules straightforwardly remove literals from a formula.

Unit propagation, though, is more complex.
Here, unit literals are removed.
Though, this doesn't necessarily remove an atom.
So, in place of atoms, a map of atoms to literal occurrences is kept.
And, when propagating, each literal from each removed set is inspected to see if any (other) instances of the literal remain in the formula.

For the moment, atoms must be inspected, as atoms removed from the formula by propagation may remain in the map.
Though, if they do, it'll be as a false / false mapping (as no positive / negative instances remain).

Note, this map was already created each instace of propagation, so even with the upkeep things likely tend to performance...