Revised unification foundations
Though, maybe to be revisited...

In place of a hashmap a list of mappings is used.
The expectation is that this simplifies mutation when solving for a fixed point.

Lookup is much less efficient, though maintaining an order for binary search through methods may help.
Otherwise, things could be revised so a hash map returns indicies to the vec.

I *think* all that's needed is some guarantee that the hashmap is disjoint, as this would, I think, allow for a mutating pass over the values of the map, while also allowing for lookups on the rest of the map.
Though, whether this works, and whether it's possible to communicate to the checker is TBD.