Term unification updates
Some dancing, though not too much.

An `update_term` method applies substitution on a term using the current env.
And, specifically, this takes a term and returns an updated term if possible, or the same term if not.
For a pass over the entire env, then, each map is inspected, the `to` term is taken from the mapping, passed through the updated method, and then returned.
This, well, is ok...