Revised literal sort ordering
Literals are now sorted by value, and then by atom.
So, false, true, and if the same value, lexicographically.

The idea is that given this method of sorting, it's 'easy' to perform operations on the positive and negative literals of a set.
For, one only needs to walk through the set until a true literal is found.

The split index *could* be cached, which avoid a vector traversal for any split operation.
Esp. as an update is required only when a true literal is added.
Still, for now, this optimisation is not made.

Given this, the previous method for filtering contradictions has been revised.
As, the method assumed literals were sorted by atom and then value.
This was nice as one needed only to compare adjacent literals.
So, no pass to split, and cache locality.
Though, generally, the revised ordering is likely to be more useful.

Elsewhere, an additional method to sort the sets of a formula set is introduced.
This helps simplify tests, and no requirement is made that sets of literals are always sorted.

And, various tests are split into more specific tests, to help track issues when they arise...