Complementary literals update, some tests
get_negative_positive_split has been updated to borrow a set, rather than relying on an index.
With this, the function has wider use, at no cost.
(And, indeed, often reduced cost as there's no need to track an index.)

Complementary literals has been updated given the new ordering, as it relied on the previous ordering as an invariant.

In addition, the function is now associated with the FormulaSet struct, to (hopefully) indicate it's reliance on the literal ordering invariant.

And, some tests have been added both for the split and for contains function.