Rough approximations toward unification of relations
I think, perhaps, the reason the book directly casts relation to functions is due to the consequences of skolemisation.
Specifically, as variables are eliminated, expected instances of unification are considered impossible.

So, there's still some work to do.

In any case, some progress is made.

unify atoms remains, though I the narrower, freshly introduced, unify_relations is all that's of interest.
Though, there's still a way to go before confirming this...

In any case, some progress...