Minor adjustments
It's been a long day, and various things aren't making sense.

Of most unsense is the Prawitz loop from the book.

I was expecting refutation by unification to roughly:

- Convert a formula to DNF.
- Search each disjunct for complementary atoms.
- If the atoms contain only constants, then the disjunct is false and can be ignored.
- Otherwise, if the atoms contain variables which can be unified, then the disjunct is false given unification.
  So, unify and ignore the disjunct.
- If all disjuncts are ignored, then the formula is UNSAT.
- Otherwise, no luck.
- Here, with caveats regarding solving the unification, maybe.

Ok, in a little more detail...

Each disjunct is given fresh variables.

Hm, some understanding?

The idea is, analogous to ground instantiation, to consider distinct variables.
P is UNSAT, only if P with all variable permutations is UNSAT.
So, similar to ground all permutations need to be considered.
The key difference with unification is that previous ...

Uff

I think the difference is that with unification information regarding relationships between varaiables is accumulated in the unifier.

Need to find some way of refuting a disjunct.
So, some tuple for which a conjunct fails.
And, tuples are important, because these roughly correspond to a ground instance.
And, it's not possible to consider a single tuple, as this may not be a refutation for some other disjunct.
In this respect, distinct variables for each disjunct could be attempted.
However, by expanding variable tuples in the same way as ground, the unifier may save significant work.
Otherwise, the information derived from unification is limited to each disjunct, as there's no way to extend the consequences of the conjuncts of some disjunct to an other disjunct.
Or, rather, doing so may overlook a refutation of the disjunct...

Well, still fuzzy, I think...

Tomorrow, perhaps...