Unify refute, revised
Together with unify complemenets, as these form a pair (at least for the moment).

Readings are quite close, as this is something of a struggle...

As I understand things, unify_refute fails as soon as some conjunction is found for which there are no complementary literals to unify.

This is because unify_complements (/ unify_literals) *does not* fail just in case a unification is found.
And, in particular, this fails immediately on a dissonant conjunction as a composite function is made from unify_refute and unify_complements.
I also expect this is the case, as without a unifier, it's not possible to establish the formula is UNSAT, and so there's little point in continuing.

unify_refute also continues immediately on finding a unifer for a complementary pair.

So, that is, immedaite continue on success for some elements of a conjunct to the next conjunct.
And, immediate exit on failure for all elements of a conjunct to the calling fn.

These are exhastive conditions, of course.

With this, the local version of unify_complements returns a boolean.
It'll return true as soon as a complementary pair is found, and false otherwise.
This, then, is all that's required for unify_refute.
As, unify_refute calls unify_complements on each conjunct, returning early (with false) just in case no unifier is found.

This seems sensible.
As, true will be returned just in case a complementary unifier is found for each conjunct of a DNF, and hence just in case the DNF must be UNSAT.

What's quie difficult, for me, is the use of failures.
As, I expect these to be important, given booeans aren't used, but they don't seem to be...