LiteralSet struct
In search of something a little mindless, a custom struct for literal sets seemed like a good idea.
As, there are a handful of ops which could be moved to methods.
And, in particular, a custom ordering which would be nice to fix structurally.

This was less mindless than expected.
Things are mostly fine, though some tidying is needed.

Specifically, literals sets are often iterated over, so there's plenty of places where the internals are accessed.

So things are still todo, such as moving setify to a method...