Set updates
After extensions `setify` is called on literal sets.
This was overlooked, and is required for the set invariant to hold.

For DNF sets, a fresh conjoin method has been defined.
This is equivalent to distributing the conjunction of two sets.
Though, general indirection for formula conjunction, or complication from additional nesting of sets is avoided.

The above led to revisting the atom cache a little.
The methods, roughly, have been extended with a `refresh` option.
And, common functionality related to updating the cache has been abstracted.
As split borrows are required things are a little messy, though function ok, I think.