Something of a checkpoint
A library for FO formulas has been added.
This contains a few Pelletier instances for now.
As, this helps avoid repetition in tests, etc.

A varson of the prawitz seems to work fine.

The implementation of unifiers needs significant revision, though.
The problem is that at present every term is stored.
And, this can led to *significant* memory pressure.
A (slow) patch to check before adding a term has been made, but it's fairly worthless.
Hm...