Unification refinements
Of a kind...

The unification map is now from variables to terms.
This is done by cloning the variables on an update, allowing a split between lookup and mut ops on the map.
There's surely something else that can be done here, but for the moment the focus is on clearing the index code.