Unification progress
Unify complements was broken.

An attempted optimisation to avoid unifying conflicting relations gone awry...

On the initial pass, pointers worked through the negative and positive literals, with both advancing on a match.
This, however, meant that unification would not be attempted for some complementary pairs.
For example, P(y), P(f_x(y)) and ~P(y_1).
Here, unification would be attempted on P(y) and P(f_x(y)) and fail due to circularity, and the unifier y_1 => f_x(y) would be skipped.

As a quick fix, all pairs are now considered during unify refute.

Outside of this fix, P45 has been updated to match the paper, almost exactly.
And, both P18 and P19 have been added.

Investigating the former helped identify the issue noted.
It was very helpful to have a simple case of this kind.