Revise internal structure of literal sets
After some detours, a key motivation for transfering literal sets to a unique struct.

Here, the set is restructured as a pair of vectors --- rather than a single vector.
These vectors contain negative and positive literals, respectively.
This avoids searching for the split, and also should generally improve setify ops, as there will typically be fewer comparisons to be made.

Todo is storing atoms, rather than literals, as with the separate sets the sign is easily found.