Minor progress
Exploring all unifications in unify retute is a task.

To avoid cloning unifiers, a trail has been added.
As an existing unifier cannot conflict with a fresh unification, unifications can be removed in reverse without issue.
In particular, this allows rolling back the freshly added unifications when attempting unify refute.

This is put to work in the new recursive implementation of unify refute.

Aside from the exploration of all possible unifications (I think), an important overlooked check was for conflicting literals given the current unifier.
Without doing this, conflicting literals were being overlooked.

Things seem ok, with the exception of P45.
This continues to be a significant issue.

So, what to do...
The current implementation is inefficient, as the same unifier will be explored, even after encountering a dead end.
Though, I wonder if there's still something I'm missing...