Some insight
A simple cache has been added to avoid exploring failed unifications.

The insight, though is returning directly from a conflict which does not require fresh unification.
As, explorining the remaining pairs in the set would not lead to a refutation.
For, if some refutation existed, it would have been explored in the (attempted) unification of some further set.

With this, stack overflow is almost immediate.
And, given the size of the sets, this is of no particular surprise.
An iterative varaint, then, before moving on...