Unify refute
Finally, I think.

The recursive to interative rewrite is mostly simple.
Comments have been added.
And, while doing so the cache has been removed.

P45 is now almost instant, which is very surprising.
So, I don't have a complete grasp of everything...
The nice thought is that previous attempts had some overlooked issue.
Though, as the rewrite should be equivalent...
Oh, well, on this though the previous *attempt* may have worked well, as at issue was stack overflow.
So, P45 being instant is a little less surprising, to the extent that I'm comfortable being more confident than not that things are okay.

As part of the above confidence some very simple sat but not taut formulas are ran through the procedure as tests, which require a limit to be exhauted.

I think it's likely time to move on...