Formula / literal sets extened with indicies.
The idea of indicies is to allow restoring a literal set to an earlier state.
This is done by storing an index in a formula set, and by marking each literal set with the index when the literal set is added to the formula (and incrementing the index).
As, so long as only additions have been made it's then possible to filter out all sets above (or below) and index.

As the indicies are additional, there is little distruption to existing code.

Significant changes are:

- A unified method for adding literal sets to formula sets.
- Updation of set creation methods (CNF / DNF) to allow use of the above method.
- A custom implementation of Eq for formula sets to ignore the counter (literal sets already had a custom impl).

Things seem to work fine.