Remove atom cache from formula sets
The cache was quite nice, and certainly helped with performance of ops such as resolution.
Still, the added overhead of maintaining the set complicated things.
Instead, a function has been addded which generates a map from atoms to number of occurrences.
This is, mostly, a drop in replacement for existing use, with only changes related from the t/f occurs predicate to occurs count required.