Resolution refinements
The occurrence map is generated for determining which atoms to inspect for blowup, and can be reused for calculating the blowup of resolution on an atom.

In principle this could habe been done before, but the implementation of the map was not sufficiently refined (or perhaps only insufficiently verified) to track occurrence *count* (and so only returned an occurrence predicate).

Elsewhere some very minor tidying.