Tableaux foundations
unify_complements has been reworked a little to deal with literal sets.
And, `pop_some` has been renamed `pop_multiple`.

No changes to existing code otherwise.

tableau diverges from the book.
Again, due to iteration replacing recursion.
Though, with an explicit struct for 'continuations' and a stack, the translation is only a little opaque.

An outer loop handles branching on disjunctions, while an inner loop handles exploring a branch.
Whenever a branch is refuted the next branch is explored by taking the top element from the stack, and refreshing the disjunct and current unification.
etc., as in the book.

Todo is handling of universal quantifiers.