Pellitier p38 fix
A second typo found in the paper, I think.

I think the intent with p38 is a biconditional between two formuas whose difference is only the names of predicates.
The lhs uses P and Q, while the rhs uses Q and P.
Though, in the paper (or at least my copy) the final occurrence of P should be a Q in order for the above to hold.

I haven't worked a counterexample by hand, but after setting up for tableau and splitting into the 32 disjuncts, it's clear the 17th (or was it 18th...?) disjunct is satisfiable.